**if statement** cool effect
added if statement so that if lineNum > 2 then the variable of lineNum
will be set to 0...this will take our event back to line 0 of the HTML
and the whole script can be run again.

This highlights each line of HTML one per click and then goes back to
top line creating an effect of changing the selection of the each lines
of code on the web page.